Google-Cloud-Functions/Writing the Cloud Function to delete VMs/requirements.txt
google-auth
google-cloud-compute